+++ src/main/java/net/c7j/weather/geomagnetic/common/ActiveType.java
@@ -1,4 +1,4 @@
-package net.c7j.weather.geomag.common;
+package net.c7j.weather.geomagnetic.common;
 
 public enum ActiveType {
