+++ src/main/java/net/c7j/weather/geomagnetic/common/ForecastIntervalType.java
@@ -1,4 +1,4 @@
-package net.c7j.weather.geomag.common;
+package net.c7j.weather.geomagnetic.common;
 
 import java.util.EnumSet;
 import java.util.Objects;
